backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
pydantic
joblib
python-jose
passlib[bcrypt]
bcrypt==4.0.1